Moved the reset_password and fotget password implementations into auth controller and auth route
