+++ backend/.gitignore
@@ -40,7 +40,7 @@ build/Release
 
 # Dependency directories
 node_modules/
-frontend/node_modules/
+frontend/
 jspm_packages/
 backend/package-lock.json
 frontend/package-lock.json
